Remove unnecessary changes made in move
