+++ samples/msal-angular-samples/angular-b2c-sample/src/app/profile/profile.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { AppModule } from '../app.module';
+import { AppModule } from './../app.module';
 
 import { ProfileComponent } from './profile.component';
